+++ doc/source/doxygen-docs/changelog.md
@@ -24,6 +24,7 @@
   - Fix bug in mrpt::math::getAngle(const TPlane&, const TPlane&).
   - Fix exception if mrpt::opengl::CFBORender is used with setProjectiveFromPinhole() camera models.
   - Fix CMake Warning at cmakemodules/FindFilesystem.cmake and failure to detect the std::filesystem feature in some g++ versions.
+  - Fix numerical innacuracies with planar bounding boxes, fixed via new `epsilon` parameter in mrpt::math::TBoundingBox::intersection()
 
 # Version 2.4.0: Released Dec 12th, 2021
 - Changes in build system: